refactor IconChooserModal
